+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### 2.x.x
 * Added a function that returns all the elements that corresponding with a tag name
+* Delete the add style option
 
 ### 2.0.3
 * Added support to input disabled